[ nicifier ] refactor: move fixity collection to a separate module
