+++ Agda.cabal
@@ -309,6 +309,7 @@ library
                     Agda.Syntax.Common
                     Agda.Syntax.Concrete.Attribute
                     Agda.Syntax.Concrete.Definitions
+                    Agda.Syntax.Concrete.Fixity
                     Agda.Syntax.Concrete.Generic
                     Agda.Syntax.Concrete.Name
                     Agda.Syntax.Concrete.Operators.Parser
